Assignments/assignment-6/requirements.txt
torch==2.0.1
    torchvision==0.15.2
    gradio==4.0.0
    Pillow==10.0.0
    wandb>=0.16.0
    huggingface_hub
    git-lfs
    